smartsaver-backend/requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
requests
yfinance
scikit-learn
numpy
pandas
